Change links to be under "shop"
