--- customer-site/lib/models/ProductModel.ts
@@ -1,14 +1,0 @@
-export type Product = {
-    _id?: string;
-    name: string;
-    slug: string;
-    image: string;
-    banner?: string;
-    price: number;
-    brand: string;
-    description: string;
-    category: string;
-    countInStock: number;
-    colors?: [];
-    sizes?: [];
-};
