--- customer-site/tailwind.config.ts
@@ -1,50 +1,0 @@
-import type { Config } from 'tailwindcss'
-
-const config: Config = {
-  content: [
-    './pages/**/*.{js,ts,jsx,tsx,mdx}',
-    './components/**/*.{js,ts,jsx,tsx,mdx}',
-    './app/**/*.{js,ts,jsx,tsx,mdx}',
-  ],
-  theme: {
-    extend: {
-      backgroundImage: {
-        'gradient-radial': 'radial-gradient(var(--tw-gradient-stops))',
-        'gradient-conic':
-          'conic-gradient(from 180deg at 50% 50%, var(--tw-gradient-stops))',
-      },
-    },
-  },
-  plugins: [require('daisyui')],
-  daisyui: {
-    themes: [
-      'light',
-      'dark',
-      {
-        cupcake: {
-          primary: '#ff6b81',
-          'primary-focus': '#ff487a',
-          'primary-content': '#ffffff',
-          secondary: '#f6d860',
-          'secondary-focus': '#f5d152',
-          'secondary-content': '#ffffff',
-          accent: '#37cdbe',
-          'accent-focus': '#2aa79b',
-          'accent-content': '#ffffff',
-          neutral: '#fbfbfb',
-          'neutral-focus': '#f5f5f5',
-          'neutral-content': '#1f2937',
-          'base-100': '#ffffff', // Light green color
-          'base-200': '#f9fafb',
-          'base-300': '#d1d5db',
-          'base-content': '#1f2937',
-          info: '#2094f3',
-          success: '#009485',
-          warning: '#ff9900',
-          error: '#ff5724',
-        },
-      },
-    ],
-  },
-}
-export default config
